(add_file): make len argument an pointer to an integer
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@12775 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ lib/krb5/context.c
@@ -271,7 +271,7 @@ krb5_set_config_files(krb5_context context, char **filenames)
 }
 
 static krb5_error_code
-add_file(char ***pfilenames, size_t *len, char *file)
+add_file(char ***pfilenames, int *len, char *file)
 {
     char **pp = *pfilenames;
     int i;
